current voting sprint data from backend

+++ SmartValley.Domain/InvestorVotes.cs
@@ -5,7 +5,7 @@ namespace SmartValley.Domain
 {
     public class InvestorVotes
     {
-        public long TokenAmount { get; set; }
+        public double TokenAmount { get; set; }
         public List<Guid> ProjectExternalIds { get; set; }
     }
 }
